requirements.txt
Package                      Version
---------------------------- -----------
alembic                      1.3.1
altair                       5.4.1
annotated-types              0.7.0
anyio                        4.6.2.post1
asgiref                      3.8.1
asyncpg                      0.30.0
attrs                        24.2.0
bcrypt                       4.3.0
blinker                      1.6.2
cachetools                   5.5.0
certifi                      2024.8.30
cffi                         1.15.1
charset-normalizer           3.4.0
click                        8.1.7
colorama                     0.4.6
cryptography                 41.0.3
distlib                      0.3.8
ecdsa                        0.19.1
fastapi                      0.115.5
filelock                     3.13.1
gitdb                        4.0.11
GitPython                    3.1.43
google-ai-generativelanguage 0.6.10
google-api-core              2.23.0
google-api-python-client     2.152.0
google-auth                  2.36.0
google-auth-httplib2         0.2.0
google-generativeai          0.8.3
googleapis-common-protos     1.66.0
greenlet                     2.0.2
grpcio                       1.67.1
grpcio-status                1.67.1
h11                          0.14.0
httpcore                     1.0.7
httplib2                     0.22.0
httpx                        0.28.1
idna                         3.10
iniconfig                    2.0.0
itsdangerous                 2.1.2
Jinja2                       3.1.2
jsonschema                   4.23.0
jsonschema-specifications    2024.10.1
jwt                          1.3.1
Mako                         1.2.4
markdown-it-py               3.0.0
MarkupSafe                   2.1.3
mdurl                        0.1.2
narwhals                     1.13.5
numpy                        2.1.3
packaging                    24.2
pandas                       2.2.3
passlib                      1.7.4
pillow                       11.0.0
pip                          25.0.1
platformdirs                 4.2.0
pluggy                       1.5.0
proto-plus                   1.25.0
protobuf                     5.28.3
psycopg2                     2.9.10
pyarrow                      18.0.0
pyasn1                       0.4.8
pyasn1_modules               0.4.1
pycparser                    2.21
pydantic                     2.9.2
pydantic_core                2.23.4
pydeck                       0.9.1
Pygments                     2.18.0
pyparsing                    3.2.0
pypdf                        5.1.0
pytest                       8.3.3
pytest-asyncio               0.26.0
python-dateutil              2.8.2
python-dotenv                1.0.0
python-editor                1.0.4
python-jose                  3.4.0
python-multipart             0.0.17
pytz                         2024.2
ratelimit                    2.2.1
redis                        5.2.1
referencing                  0.35.1
requests                     2.32.3
rich                         13.9.4
rpds-py                      0.21.0
rsa                          4.9
setuptools                   65.5.0
six                          1.16.0
smmap                        5.0.1
sniffio                      1.3.1
SQLAlchemy                   2.0.20
sqlparse                     0.5.3
starlette                    0.41.2
tenacity                     9.0.0
toml                         0.10.2
tornado                      6.4.1
tqdm                         4.67.0
typing_extensions            4.12.2
tzdata                       2024.2
uritemplate                  4.1.1
urllib3                      2.2.3
uvicorn                      0.32.0
virtualenv                   20.25.0
watchdog                     6.0.0
Werkzeug                     2.3.7
